fixed mixeval edit, copy, and delete
When a mixeval is editted without any questions that have descriptors,
an invalid index error will occur. I've fixed that by only processing
descriptors if any exists. Another problem still exists when additional
questions are added when an a validation error occurs. When that happens
any new questions will disappear. We're still investigating this
problem.

I've fixed delete where it had an error of using a function that
was in evaluation_base. I've modified it to use it's own function.

I've modified copy again to make it similar to modifying simple
evaluations and rubric evaluations, where all of the parent
template's detail is copied onto an "add form".