included a filter for rubric evaluation results
The filter for both the student and instructor view of rubric
results is to not include evaluation_rubric scores when the
evaluation has not been submitted. This was also done in mix
evaluation earlier, however this will not been done for simple
evaluation because evaluation simple entry is not saved
separately from evaluation_submission.