fixed review status for evaluation results
Evaluations are set to have a "to review" status when everyone
in the group has submitted, but all three evaluation types have
some bugs in the logic for determing when to set the "to review"
status. Some bugs include not using the correct number of group
members. Instead of using the number of group members from the
groups_members table, they were using the number of evaluatees
in the form, which is incorrect.

I've also changed the query to find out the number of submissions
in the group, which is the additional condition that the submitter
must be in the list of group members. In other words not include
submitters that are no longer in the group.