included an additional instruction to simple, rubric, and mixeval
The additional instruction that is being added to the students'
evaluation forms is the note that the evaluation can be
repeatedly submitted until the release_date_end. The reason
for the note is that students no longer have to contact their
instructor to reopen the evaluation for them if they want to
resubmit. The feature was implemented for version 3.0