initial commit for permissions editor
This feature is based on Ticket #429. Super admins can view
the permissions for each individual role at accessses/view/$roleId.
If no $roleId is given, the default role is super admin. The user
can change the role by using the drop down menu near the top of the
page.

By clicking on the specific access, you can choose to allow or deny
ALL actions (create, update, read, delete) for the access OR allow
or deny the access for a specific action.