finished implementing the permissions editor
I've added some parameter checks to stop the function from proceeding
if any of the parameters are invalid such as a 6 for role id
or an invalid acos id.

I've also put the link to the editor on the admin page.

Furthermore I have, created test cases for the access model and
controller.