fixed simple eval view and modified rubrics edit
Simple eval view uses the same element as the actual simple
evaluation form the student fills out including the instructions
at the top. The instruction that was added earlier which tells
the students when the evaluation stops being accessible caused
a bug because the preview does not have a date to display. I've
fixed it by replace the date with a string that says Evaluation's
release end date.

I've also removed the button linked to the edit view for rubrics
to keep it consistent with the other evaluations. Simple and mix
evaluations and surveys view do not have links to the edit page.
Please enter the commit message for your changes. Lines starting