Ticket #495
I have added a  check for the dom extension and DOMDocument class
during the installation process and when a user tries to access
TeamMaker.

DOMDocument is now part of the list of optional extensions. There
is also a help text that tells the user that it is required for
TeamMaker but optional for iPeer.

When a user tries to access TeamMaker and the dom extension and
DOMDocument class cannot be found a flash message will appear
to tell the user to contact the system administrator to use
TeamMaker.