added survey validation and modifies events add/edit
I've included a validation for survey to check the availability
field. If it is left unselected, a validation error will now occur.

I've added two buttons to events add/edit to select or unselect
all the groups. This is to make it more convenient for instructors
that just want to select all the groups.