removed record_status from the groups add/edit form
Record status is no longer used in the application, so I've removed
it from the form so that the users can't modify it.

I've also refactored simple eval form a little bit by modifying
a permission check (whether the user belongs in the group) from
using two queries and a loop to one query.
ease enter the commit message for your changes. Lines starting