created a new system test for adding instructors
I've created this test to especially test the new instructor form
field. I've also tested deleting the course (with the url) so that
we can use the test many times (courses cannot have the same name).
I could not test the delete functionality with the action link in
the menu generated by ajax_list

I've also changed the $wd_host to use localhost instead of the ip
address so that the ip address will not be hardcoded into the code
and required to be changed whenever a new user  uses it.