fixed mixeval zero mark bug
When the zero mark option is selected in the mixeval template,
the weight of each descriptor is shown correctly in the add and
edit view. However when we view the template in the view and
student mixeval_form, the zero_mark selection is ignored.

I've patched it up by modifying the conditions for the zero_mark
option to be used.