fixed the import features that require uploading a file
The path to the temporary file became incorrect when we were
uploading the file on the ipeerdev server. This change has been
made to users import already. This commit includes the change to the
groups import and course import (mass move)

In this commit, I have also made the webpage link on the courses/home
page to open a new tab/window instead of using the current tab.
This is to make it the same as opening the course homepage
from the courses/view view.