Added the ability to export SimpleEvaluations(only) as pdf
Modified evaluations_controller.php and export.ctp to add
the option to export SimpleEvaluations as pdf's. The options
selected on that screen (for example, Include Coursename) do
not affect the PDF output. I am planning to take those params
into consideration in the next commit.
Added new file export_pdf.php which makes calls to the
XTCPDF wrapper class in order to generate the PDF output.