fixed events add/edit javascript
When the user switches the event tempalate type to survey the
email reminder frequency field disappears. The reason was the
field uses radio buttons. All radio button fields disappear when
the template type is changed to survey.

I've fixed it by adding an id to the email reminder field to
differentiate between the other radio fields and it. Now the field
is shown at all times.