fixed emailer ajax search
The emailer only has one field that we can search by entering
a value in the textfield, which is the subject field. However
when we enter any partial parts of the subject line to the field
no email schedules are returned. When we enter an id (eg. 1),
however, we get one email schedule returned.

The reason that the emailer searh is using the email schedule id
instead of the subject field was because the searchBy value set
in the emailer controller was set to the id instead of subject.
This has now been changed.