Rubrics can now be exported as pdf's
Modified export_pdf.php and added the ability to export rubric
evaluations as pdf. The options selected on the export page
are not taken into consideration as of now and the pdf is similar
to the results page for rubric evaluations(detail). Refactoring needed
in export_pdf.php (which I will do before my next commit).