modified ajaxlist mapping
I have removed the addition of 's' or 'es' to headings of fields
that we are mapping, because it has started to created incorrect
words such as Availabilitys. Having the words in their singular form
seems to make more sense.

I have also mapped all availability fields for all evaluation types
and email templates so that users do not need to type in private
or public in the search field.

I have also fixed a bug in the groups import where students that are
in the class are unable to be imported. The reason was when we merged
the tutors and students in the class, if the user_enrol and user_tutor
id matches, one will overwrite the other. This bug has been fixed.