made some minor changes to severy views
Removed link to iPeer home page in edit profile successful message
because we no longer required users to enter to enter a email
in their profile.

Added a space beforethe word comments in the simple eval form
so that the word is separate from the number.

Changed several views to show the student no field instead of the
title field for tutors

Added a javascript confirm for resetting passwords in the class
list view