included more tests for events and groups + mixeval fix
More tests are written for groups and events to test out error
checking and some corner cases.

We have also fixed mixeval's questions editor. The original fix
for a bug that occurs for a corner case, which is when the template
initially only has one question, introduced another bug. The new
bug is when all questions are cleared and a new question is added
the first question is numbered 2 instead of 1. This has now been
fixed by changing the way the array of question numbers is created.