created tests for rubrics events
The tests include students/tutors answering the evaluation and
evaluation results. I have also fixed penalty for student results.
The final penalty is not implemented even if the current time is
past when the release date ends. I have also added a relationship
between events and evaluation rubric so that when a rubric event
is deleted, the related evaluation rubric entries would also be
deleted.

For the instructor's view of the evaluation results (for rubrics)
had a bug for the string that tells whether the student is
above or below the group average. The average was initialized to
zero and not calculated afterwards. This has now been fixed.

For student simple system tests, I have added tests for basic view
of student view results and tests to make sure tutors are never
evaluatees.