added some system tests for edge cases
Some system tests have been added to add_simple and student_simple
for testing that negative numbers are not valid. I have also fixed
a bug in simple evaluations for students where they can give group
members negative points.

I have added some system tests to all the import features which
include users import, groups import, and mass move to test that
importing with files of invalid file types will not go through.
I have fixed a bug for groups import and mass move where the user
is redirected to a blank page if a file that is not in a csv or
txt format is imported.