fixed a bug in the api and rubrics results
Detailed rubric results have an invalid index error that has now
been fixed.

The api has a small bug for enrolment where the classlist may not
be complete if any of the ids in user_tutors, user_courses, and
user_enrols tables are the same. This has now been fixed by
using array_merge instead of the union operator (+).