modified evaluation email reminder scheduler
The changes made include the following:
- start date is the later date between evaluation release start date
and current date. The reason is the email scheduling may happen
before or AFTER the evaluation has already started. If it is after,
students may get multiple emails in the beginning which were all
scheduled between the start date and current date.

- when the reminder frequency is changed, only the emails that have
not been sent would be deleted.

- email reminders can now be set for surveys. Previously this could not
be done due to survey events do not have groups

- emailer schedule to field is updated when the email is sent only with
the students that haven't submitted and are still within the group
or class(survey). For example, unenrolled students would not get
emails to take a survey even if they were in the original list of
recipients. The vice versa will also happen. If a new student is enrolled,
they will still get the email reminders if they haven't submitted.

- frequency calculations have also been changed. If two or more emails
are scheduled, the frequency becomes the difference between the last
two. If only one email is scheduled, the frequency is defaulted to 7 days.

I have also changed the emailer to send the emails as text, so that the
url given in the email content will become links.