fixed evaluation results errors and fixed accordion's cursors
I have fixed a bug where there invalid indices and variables exist.
This has now been fixed for all of the instructor's evaluation
results views.

I have also made the cursor to be a pointer when the mouse hovers
over the header of the accordions found in the instructor's
rubric and mixeval results view, student's rubric results view,
and the rubric evaluation form. This is to make it more intuitive
than the default "text" cursor, which was pointed out by one of
our users.