made minor modifications throughout the application
I have refactored the text layout for the addUser notification
email. I have removed all html tags which just gets printed as
plain text and rewrote the message to be similar to the html
layout.

I have fixed some grammar problems and reworded some help text
and instructions throughout the application.

Lastly, I have fixed the css for rubrics and simple evaluation
preview to use a pointer to toggle instead of the auto cursor,
which is not intuitive.