modified email reminders to include the link to iPeer
The email reminders for events now have a link to iPeer. The shell
cannot access Router therefore we had to access the information via
the database. Therefore we have now saved the base url as a sys
parameter, which is added automatically during installation and
during upgrade. However if a developer decides to drop the database
and import one of the sql files, the base url parameter is blank.
They will have to add the url themselves