fixed a bug for adding/editing courses
If an instructor was created in version 2.x, they would not belong
to any departments or faculties. This leads to them not being able
to add their course to any departments or add instructors to their
courses.

The problem with adding to departments was fixed previously by
allowing the instructor add their course to any of the departments
in the system. However the bug still exists for instructors.

I have fixed the probem with invalid indices when an instructor
goes to courses/edit with instructor(s) that would not appear in
the user's original list of instructors. One example would be an
admin adding an instructor from a different department to the course.
This has now been fixed.

A similar solution is also done for courses/add. The instructor would
be able to add any instructors from the system to their course.