fixed emailer view
The emailer view was not displaying the email content of the
event reminders because the content is not saved in the database.
Instead it is rendered with an element. To fix this I echoed
the html format of the content instead of using the text area.

Another problem was the line breaks were not displaying correctly
even though they are displaying correctly in the actual email. I
have fixed this by replacing the \n with <br>.

I have also changed the permissions mechanism for the emailer view
and index. The instructors are given access to all emails created by
themselves and instructors that they share courses with.

Lastly, I have modified the add_event system tests to include the
changes I made.