modified courses import to be more flexible
The feature now allows moving students from courses that do not have
survey events. There is now also a check and notice to the instructor,
when a student is moved and has already submitted a peer evaluation.
Another modification made is students in the files do not have to
be in the source course. They could be new students.