added more tests for moving a group of students from a course
Some of the tests I have added includes the changes I have made
to the features, which includes allowing students on the list
to not be in the source course, allowing moving students from
courses that does not have a survey event, and the new message
which notes whether a student has submitted to a peer evaluation
or not.

I have also added a new column to the events index which is the
number of groups in the event. If the event is a survey, it will
say N/A since surveys do not have groups. If the event has no groups
a red "0" will be given to warn the instructor that no groups are
added.