modified groupMembers and enrolment in the api
For groupMembers, I have changed the reaction to adding a group
member that is not in the course to skipping the user from terminating
the request. I have also added the task of removing members that
are not in the new list.

for enrolment, I have changed the reaction to enroling users that
do not exist to skipping the user from terminating the request.
Similar to groupMembers, users already enroled in the course but do
not appear in the new list would be removed.

I have added tests for the api to test the changes mentioned above.

Conflicts:

	app/controllers/v1_controller.php
	app/tests/cases/controllers/v1_controller.test.php