added Google Analytics to iPeer
The sql files have been updated with two new fields which are
the tracking id and the application's domain (Google Analytics Beta).
The Google Analytics script is only generated if a tracking id is
given which can be set in System Parameters by the system admin. It
would be queried in the app_controller.

The Classic script is generated if only the tracking id provided. If
a domain id is also provided, the Beta script would be generated.

On a side note, the system parameter forms now trims the fields
before saving, which can help avoid problems with extra spaces
before and after the text. Validation has also been added to
system parameter to make sure the parameter code is unique. Lastly,
the system parameter controller tests have been updated to reflected
the newly added system parameters.