modified submission reminders for rubric
Previously, the reminder students get when they leave the rubric
evaluation form without submitting will get an alert reminder. They
also get a message when a evaluatee's section is saved that shows up
for a very short period of time. I have modified the evaluation
reminder to be an alert message on the top of the form every time
a section is saved. The reminder also points out comments may be
missing if they are required and a comment section is left unfilled.