modified rubric evaluation form to not have default Likert answers
The rubric evaluation form has been modified to not have the lowest
rating set as default for the Likert question. Due to this change
all likert questions have to be answered before saving. A javascript
validation has been added. If a Likert question is left unanswered,
the question would be marked red and a warning message will appear
below the save button telling the student to answer all the questions
marked red before saving.