fixed installation in pagodabox.
The application was not recognizing the database is already predefined.
The reason was the bootstrap was running before the database.php.
This has now been fixed.

I have also fixed a minor bug in the last step of the install
process where a printf broke and caused the view to not be fully
rendered.