fixed broken tests
The tests were broken because they were coincidentally ran 2 days
after one of the events in the sample data. I have now fixed the
problem by modifying the sample data events' due dates to be more
flexible. The "year" is now set to the next year (eg. now is 2013,
the date is set to 2014)