modified all the system tests to make them more reliable
The system tests were sometimes failing because the browser was
loading the next view too slowly. Furthermore, the tests that fail
if the tests fail vary from time to time. To help make the tests
more reliable, I have extended the WebDriver's element and
elements functions to include a wait for the element to appear first
before proceeding. This will accomodate for the times the browser is
slow.