refactored the application's header from the original blue header
I have placed the iPeer logo on the left. On the right the super
administrator has the option of adding a "custom" logo such
as the institution's logo. The recommended height for the custom logo
is 40px.

To use a custom logo, the logo has to be placed in webroot/img/layout/
within the app folder. Then the super administrator has to go to
the system parameter index and edit the banner.custom_logo parameter's
value to the img's name such as ubc.png.