added a new system parameter for setting the timezone
A new system parameter has been added to the system to allow the
super administrator to change the timezone. The timezone is initialized
during the installation process. If the super administrator would like
to change the timezone, they can go to system parameters under the Admin
tab and edit the system.timezone parameter. If an invalid timezone
is provided, the system defaults to UTC and a warning message will
appear warning the super administrator that an invalid time zone
is provided.