minor modifications including changing the list of valid timezones
I have removed the function that generates the list of valid
timezones because it only includes the current timezones and
not the backward compatible timezones. I have changed it to use
listIdentifiers instead.

I have fixed a bug for exporting simple evaluations to csv where
the evaluator's penalty is applied to their evaluatees' grades.

I have also fixed a bug where members that have not submitted
their evaluations but "now" is between due date and release end
date getting the final penalty instead of the expected no penalty.

Furthermore, I have removed the use of NOW() in the event model and
replaced them with date, so that the timezone set in system
parameter is used instead of the machine's timezone.