modified some system tests to reflect the removal of a button
I have removed the "Now" button from the calendar pop-up previously,
so I have now updated the system tests to not use the button.

When no timezone is set in the php.ini file the a notice about an
unreliable source of the timezone will appear in test.php. This has
now been fixed by setting the timezone to UTC