fixed a bug in evaluation results for rubrics and mixeval.
For some odd reasons, sometimes a user in the system who is not in
the group will appear on the not submitted or not in group list if
they were suppose to be empty. The strangest part of this bug is
the algorithm that generates the list of user ids returns empty
(correct) but the when we put the empty array into the getUsers
function, it returns a user. I fixed the bug by checking whether
the userIds is empty. If it is, the function will just return an
empty array.

I have also fixed some controller and model tests that were missing
fixtures.