fixed broken tests
The tests were broken because the due dates set in the sample data
has past which causes the events to be overdue, which was not what
we expected in tests. The sample data now sets the due dates to the
next year. Some of the tests have also been modifed to accomodate
for the more "flexible" due dates.