allow time zone to be set in php.ini.
Previousy when we implemented time zone settings, we assumed no
time zone is set in php.ini. However many systems already has their
timezone set in the php.ini file especially if the administrator
is just upgrading. I have now modified the code to first check
whether a time zone is already set in the php.ini. If it is, that
time zone would be used.