Modified mixeval add/edit/copy to have to a self-eval section
This is the initial commit for Ticket #502. A new self-evaluation
section has been added to mixed evaluation. The questions in this
section are only for the evaluator.

The question sections were moved to the questions editor instead of
staying in the add view because we would need to echo the questions
editor element twice which will duplicate the functions.

I have modified the sql files to include a self_eval field
for the mixeval questions table to differentiate questions between
the two sections. I have also modified the numbering algorithm for
the questions so that the questions in the two sections are not
numbered together (eg. peer eval has 3 ques, self eval's first
question will be numbered 1 not 4). However behind the scenes the
questions from the two sections are numbered together (eg. from
the example above, the first self-eval question is numbered 4 in
the table).

I have also edited some model and controller tests to reflect the
changes made to mixed evaluation.