Included self-eval to mixed evaluation results
In the instructor's detailed evaluation results view, an
accordion for self-evaluation will be created if any self-evaluation
questions are included in the evaluation. The results
are grouped by questions to allow easy comparisons.

In the instructor's basic evaluation results view, a table
for self-evaluation will be created if any self-evaluation Likert
questions are included in the evaluation. The grades displayed are
the averages of each question. This allow instructors to know
on average what the members gave themselves for the questions. Also
the question average is more useful than the sum of the marks the
students gave themselves.

In the students' detailed evaluation results view, a self-evaluation
section will be created if any self-evaluation questions are included
in the evaluation.

In the students' basic evaluation results view, a self-evaluation
section will be created if any self-evaluation Likert questions
are included.

Overall, I have also modified some of the elements to allow
customizations to the text to reflect the difference between
peer-evaluation and self-evaluation questions. Also I have added
a check whether the evaluation has peer-evaluation quesitons. This
is to introduce more flexibility where we can have evaluations
that only have self-evaluation questions.