Modified evaluation templates' views' headers
The different evaluation templates' view, add, edit, and copy
headers have different formats. Therefore, I decided to make
them all the same. For add and copy placed Evaluation Type >
Action, which the Evaluation Type being a link to the index. For
edit and view, Evaluation Type > Action > Name.

I have also modified some wordings. For a column in evaltools
index (mixeval), I have changed reflective questions to self-
evaluation questions so that it would be the same as the rest
of the application. I have also modified the help-text for
self-eval in mixeval add/edit/copy.