Modified courses/edit for instructor in no faculties
When instructors are created with no faculties, their list of
available instructors is empty. This causes a problem when the
course is edited (Eg. change the title). Since the list of instructors
would be empty, the list of selected instructors would also be empty.
This will cause the list of instructors in the course to be removed.

I have now fixed the problem by appending the course's instructors
to the beginning of the list of available instructors whether
the instructors are in the same faculties or not.