Added a small event model test
I have added a small test for event model to test the due_in
attribute. The test checks that the model uses the php time zone
instead of the sql time zone when calculating the due_in field.
The test uses assertWithinMargin, because the difference between
the time within the test and time used by the model may be
off by a few seconds depending on the speed of the machine. The
difference is alright since we mostly care that the correct time
zone is used which has differences measure in hours.

The test will have to be modified in version 3.1 to accomodate for
the possibility that the timezone is set in system parameters and
not in php.ini.