Fixed courses edit
A bug was found when user A edits the user B's profile if User A does
not have access to all of the courses user B is enrolled in. The problem
was user B would be unenrolled from the courses User A does not have
access to.

I have also copied the fix from dev for another problem which is
if all the courses are unselected, the user would not be unenrolled
from any of the courses.