Fixed reported bug in student index
The bug is events that are not released (release dates are in the
future) are placed in the expired section, when they should not be
displayed at all. The reason behind the bug is events in the above
situation are fulfills one of the filters requirements (no submissions
and date is before result_release_date_end.

To fix this bug I have added an extra condition which is the current
date and time has to also be after release_date_end for the event
to be placed in the expired section.

I have added an extra event to test this bug. The event is set to
be released "next" year on July 31st. Using the existing tests, we
can check that the bug is permanently fixed. The new event should not
appear in the expired array or any other events' arrays.