Added Criteria View Mode
Implemented a feature to display rubrics sorted by criteria as opposed to
being sorted by students.  Instructors will be able to select the view
mode when creating their rubric.  When selecting the view, instructors
will be shown an example of what the rubric will look like in either mode.  This feature only affects the student's view of the rubric, not the instructor's view.

This feature also required an upgrade to the database as it introduced a
new storage value for the view.  This value could either be: 'criteria' or
'student'.