Updated Test Cases and Read-Only Student Number
This commit updates the controller test cases. Not all the methods have
working test cases, but the ones that failed were fixed.

In addition, this commit includes minor bug fixes that appeared after
commit 7900750f62 (i.e. Ticket #361). The bugs include:
- Error message appearing when viewing rubrics because $studentId was undefined
- Error screen when submitting mixed evals

This commit also includes code to wrap textual survey responses.  Before
if the text was too long it would continue off the page.

Finally, this commit adds a system parameter that enables students to be
able to change their student number or not.  The parameter can either be
true/false and the default value is true (i.e. students can change their
number).