Text Wrap for Survey Response and Timezone Fix
In this hotfix, a minor change was made to the multi-line survey
responses.  Before, the text would continue off the screen for as long as
the user's response.  Now, the text wraps around and stays within the
bounds of the screen.

In addition, the timezone fix added in the dev branch (commit
7825c55a6c4fe86dae93aa8ca2aac61dcc49f02b) was added here as the model test
case failed otherwise.  The fix checks to see if the user is in daylight
savings time or not.  If the user is not in daylight savings time, it adds
an hour (3600 seconds).