Evaluation Submission Bug Fixes
This commit fixes a bug in Simple Evaluations.  When allocating points in
Simple Evals, the JavaScript should prevent the user from submitting if
the total allocated points is greater than the required number of points.
Now, when the user allocates more than the total number of points, the
JavaScript will indicate that the user needs to deallocate 'X'number of
points and will disable the submit button.

In the 'mixeval_questions' table, when creating a Mixed Eval, nothing was
being saved to the 'scale_level' column.  This commit ensures that the
correct value is saved in the 'scale_level' column.

In addition, this commit also adds code that will prevent Mixed Evals and
Simple Evals from being saved if the user removes the HTML 'disabled' part of
the Submit button.  I also added code that will save the correct value of
Mixed Eval and Rubric questions even if the user happened to change the
HTML value.  Before users could potentially change the value to a large
number and it would be saved to the database.

Finally this commit, adds the missing test groups when running the "phing
test-win" command (Windows).  There are also a few updates to the
Departments and Faculties Controller test cases.