Change OAuth URL to use set parameter for signature
When the web node behind the load balancer, who is off loading SSL, the
URL Route::url from CakePHP will only get http as protocol instead of
https. This will cause "invalid signature" error as the base string for
calculating the signature are different. So the workaround is to set the
application URL in the system parameter and this patch will use it
instead of getting the domain from Route::url.