fixed timezone test
I have fixed/modified timezone test cases in the event model. Firstly,
I have increased the fudge factor to 10 seconds. The reason is sometimes
the computer may be slightly slower and get a result of 6 or 7 second
difference which causes the tests to fail. Furthermore I have modified
the test cases to use an updated $dueIn value, or else the three test
cases are comparing the same values three times and not have the
opportunity for the code to change the timezone.

Note: the three $dueIn values for the test cases will be expected
to be the same because the amount of time left to complete the
evaluation should stay the same no matter which timezone we are in.