fixed the inability to view students' survey
Previously if you are an administrator and have access to the course,
the system still returns an error message that it is an invalid id
when you are trying to access a survey. The reason is the system only
checks whether you are an instructor of the course or not. To fix this
problem, I have changed it to check whether the logged in user has
access to the course if they are trying to access the surveys not as
a student.

Conflicts:

	app/controllers/evaluations_controller.php