Fixed problem with exporting evaluation results to csv
I have discovered that when we export all results from evaluation
events that do not have the self evaluation option turned on, we
still receive rows of empty result for self evaluation.

For example, if we have no self-evaluation event and we exported all
the results we still get a row that have no results (as expected) for
Ed Student from Ed Student.