Vagrant setup + update readme
I have added installation of the oauth php extension and other plugins
to allow tests to work. Currently the majority of the tests work on
both the browser and terminal (phing) except for v1_controller which
only works on the terminal. I will be working on a solutin for the tests
to run on the browser.

I have also updated readme with instructions to create and use the
vagrant virtual environment.