fixed bugs in saving mixeval evaluation details
There were two bugs when saving mixeval evaluation details. The first
one is trying to answer a likert question with a level of mastery of
zero. If zero is selected, the answer would not be saved. The reason
was likert questions answers are checked for validity before proceeding
by checking whether the "selected_lom" attribute is "empty". However in
php empty( 0 ) returns true, therefore the answer is ignored. This has now
been fixed.

Secondly, events that use mixeval templates that set the lowest level
of mastery to zero marks do not save correctly. The grade saved
does not take into account the lowest LOM is zero. In other words,
it is calculated as if the "zero_mark" attribute is false.

I have also update the readme documentation regarding Vagrant.