Fix #550, added ability to use email templates for event reminders
Firstly, I have added four new email merges which includes course
name, event title, event due date, and event close date. This allows
users to create email templates that are more useful as a event
reminder.

The SendEmails Shell has been modified to merge the new merge keys,
since the original algorithm assumes the id is a user id. Currently,
only if a event id and a course id is saved with the email schedule,
would the merger work. If the ids are not present, the merge keys would
be ignored during the merging process. In other words, if an instructor
sends out a email via iPeer (not throught event creation) and using a
email template that has the event's due date key, the key will stay as
{{{EVENTDUEDATE}}} in the message.

In events add/edit view, the list of email templates is not visible
until a frequency rate of more than zero (disable) is chosen.

To be able to include the ability to edit the event reminders in
events/edit, including the template used, I've saved the email
template id in the email schedule's content field. When the actual
subject line and content is needed in the sendEmail shell, the content
would be retrieved from email_templates.